+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### NEXT_VERSION_TYPE=MAJOR
 ### NEXT_VERSION_DESCRIPTION_BEGIN
 * **breaking changes** yamoney и yandex money заменены на yoomoney
+* удалена зависимость от yamoney-git-client
 ### NEXT_VERSION_DESCRIPTION_END
 ## [5.2.2]() (30-11-2020)